erts: Fix db_match flatmap copy bug
If the body of a matchspec would return a flatmap with
a variable ('$1', '$_' etc) as one of the keys and the
variable was not an immidiate, the key term would not
be copied to the receiving processes heap. This would
later corrupt the term in the table as the GC could
place move markers in it.

Also fixed a bug in the stack estimation logic when
a flatmap with all constant values, but not constant
keys was encountered.

Closes #7683